+++ src/constants/draw.ts
@@ -1,2 +1,0 @@
-export const FRAMERATE_DRAW = 1000 / 60
-export const FRAMERATE_SELECTION = 1000 / 60